requirements.txt
streamlit>=1.28.0
openai>=1.0.0
elevenlabs>=0.2.0
pyttsx3>=2.90
moviepy>=1.0.3
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
pathlib2>=2.3.7 